+++ src/main/java/pl/pawc/ewi/model/KilometryRaport.java
@@ -1,9 +1,11 @@
 package pl.pawc.ewi.model;
 
+import java.math.BigDecimal;
+
 public interface KilometryRaport {
 
     String getMaszynanazwa();
     String getMaszynaid();
-    Double getStanpoczatkowy();
+    BigDecimal getStanpoczatkowy();
 
 }
