+++ src/main/java/pl/pawc/ewi/model/RaportStan.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -12,8 +14,8 @@ public class RaportStan {
     private String maszynanazwa;
     private String maszynaid;
     private String jednostka;
-    private double normaid;
-    private double stanid;
-    private double stanpoczatkowy;
+    private long normaid;
+    private long stanid;
+    private BigDecimal stanpoczatkowy;
 
 }
